♻️ [refactor] #110 장면추가 툴팁 별도 파일로 분리
